some refactoring and output improvements
