+++ src/native_main/main.cpp
@@ -5,8 +5,14 @@
 #include <MockBarometer.h>
 #include <MockIMU.h>
 #include <State/State.h>
-#include "AvionicsState.h"
-#include "AvionicsKF.h"
+#include <State/AvionicsState.h>
+#include <Filters/AvionicsKF.h>
+
+template <int N> void printVec(mmfs::Vector<N> vec) {
+    for (int i = 0; i < N; i++) {
+        printf("%7.2f ", vec[i]);
+    }
+}
 
 int main()
 {
@@ -22,6 +28,7 @@ int main()
         "_", "_", "_"
     };
 
+    mmfs::Logger logger;
     MockBarometer baro(dataPath, "B-Pres (hPa)", "B-Temp (C)");
     MockGPS gps(dataPath, "G-Lat (deg)", "G-Lon (deg)", "G-Alt (m)", "_", "G-# of Sats");
     MockIMU imu(dataPath, accColNames, gyroColNames, magColNames);
@@ -34,12 +41,23 @@ int main()
 
     mmfs::AvionicsKF kf = mmfs::AvionicsKF();
     mmfs::AvionicsState avState = mmfs::AvionicsState(sensors, 3, &kf, true);
+    logger.init(&avState);
     avState.init();
+
+    printf("%7s %7s %7s %7s %7s %7s %7s %7s %7s\n",
+        "PX", "PY", "PZ",
+        "VX", "VY", "VZ",
+        "AX", "AY", "AZ");
 
     while(baro.isInitialized() && gps.isInitialized() && imu.isInitialized()) {
         avState.updateState();
-        std::cout << "accZ: " << avState.getAcceleration().z() << " (m/s/s)\t|" << "vz: " << avState.getVelocity().z() << " (m/s)\t|" << " pz: " << avState.getPosition().z() << std::endl;
+        printVec<3>(avState.getPosition());
+        printVec<3>(avState.getAcceleration());
+        printVec<3>(avState.getAcceleration());
+        printf("\n");
     }
+
+    printf("Flight data finished!\n");
 
     return 0;
 }
